pyproject.toml
[project]
name = "actionkit-cli"
version = "0.1.0"
description = "Command-line interface for the ActionKit API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=26.3.1",
]

[project.scripts]
actionkit = "actionkit_cli.cli:cli"
